Added check for vimrc file before invoking mv

+++ setup.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 mkdir -p ~/.vim/bundle
 git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+if [ -f ~/.vimrc ]; then
 mv ~/.vimrc ~/.vimrc_bkp
+fi
 cp vimrc ~/.vimrc
 vim +PluginInstall +qall
